Refactor examples and ignore virtual environments
