+++ examples/benchmarks/analytic_plasma_expansion.py
